RecentChooser stuff working quite well
Requires access to the mime database. May have to fool around on windows
and define the appropriate XDG environment variable to get it working.
Will also have to copy quite a lot of files for this

Problems with mac integration: list always returns the first element.
Suspected bug in mac integration