improved escape ratios procedure
Starting with this commit, we transition to a combined variance
reduction -- brute force approach.
The variance reduction will be used for the steplength determination of
the interaction, as well as for the fluorescence yield of the X-ray
fluorescence processes (note: all PE interactions lead to fluorescence
because of this!).
After the first interaction, the brute force steplength determination
will be used to see if the photon ends up outside the detector.

There was also some code cleanup: I removed the useless
--enable-optimizations option