html xslt transformation
Since the HTML5 validator kept throwing 1000's of errors when parsing
the html files that are produced by xmso2htm, though they display correctly,
an investigation was conducted.
The main problem is that libxml2 and libxslt DO NOT support HTML5 at this point,
and instead produce an XHTML file, which leads to some tags being unnecessarily closed,
and a load of other trouble.

It should be noted that the produced file contains some real errors, like using attributes
that were deprecated in HTML5 but these are unlikely to cause a lot of trouble. Worse are the use of wrong tags...
Fortunately browsers are very, very error tolerant...