xmi_main_msim: ensure all requested photons are simulated
Until now, due to the way OpenMP is used in this method, a number of
photons will not be simulated if the remainder of the long division of
the number of photons per line or interval by the number of threads was
not zero. This patch adds the remainder photons to the master thread.
This is particularly obvious if the number of photons per line or
interval is very small.

The progress display has been updated accordingly.

Many thanks to Dima Docenko for the extensive bugreport after observing
this behavior.